@@ -13,7 +13,7 @@
 		"table",
 		"semantic-ui"
 	],
-	"version": "0.3.9",
+	"version": "1.0.2",
 	"private": false,
 	"main": "dist/index.js",
 	"module": "dist/index.js",
@@ -28,13 +28,13 @@
 	"dependencies": {
 		"@babel/polyfill": "^7.12.1",
 		"cleave.js": "^1.6.0",
-		"moment": "^2.29.1",
-		"node-sass": "^6.0.1",
-		"react": "^17.0.2",
-		"react-csv": "^2.0.3",
-		"react-dom": "^17.0.2",
-		"react-scripts": "4.0.2",
-		"semantic-ui-react": "^2.0.3"
+		"moment": "^2.29.3",
+		"node-sass": "^7.0.1",
+		"react": "^18.2.0",
+		"react-csv": "^2.2.2",
+		"react-dom": "^18.2.0",
+		"react-scripts": "^5.0.1",
+		"semantic-ui-react": "^2.1.3"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -61,9 +61,9 @@
 		]
 	},
 	"devDependencies": {
-		"@babel/cli": "^7.17.6",
-		"@babel/core": "^7.17.8",
-		"@babel/preset-env": "^7.16.11",
-		"@faker-js/faker": "^6.1.2"
+		"@babel/cli": "^7.18.6",
+		"@babel/core": "^7.18.6",
+		"@babel/preset-env": "^7.18.6",
+		"@faker-js/faker": "^7.3.0"
 	}
 }
